Module Tezos_legacy_store.Legacy_chain

Tezos Shell Module - Managing the current head.

val genesis : Legacy_state.Chain.t -> Legacy_state.Block.t Lwt.t

The genesis block of the chain. On a test chain, the test protocol has been promoted as "main" protocol.

val head : Legacy_state.Chain.t -> Legacy_state.Block.t Lwt.t

The current head of the chain.

val locator : Legacy_state.Chain.t -> Tezos_base.Block_locator.seed -> Tezos_base.Block_locator.t Lwt.t

type data = {

current_head : Legacy_state.Block.t;

current_mempool : Tezos_base.Mempool.t;

live_blocks : Tezos_crypto.Block_hash.Set.t;

live_operations : Tezos_crypto.Operation_hash.Set.t;

test_chain : Tezos_crypto.Chain_id.t option;

save_point : Stdlib.Int32.t * Tezos_crypto.Block_hash.t;

caboose : Stdlib.Int32.t * Tezos_crypto.Block_hash.t;

}

All the available chain data.

val data : Legacy_state.Chain.t -> data Lwt.t

Reading atomically all the chain data.

val known_heads : Legacy_state.Chain.t -> Legacy_state.Block.t list Lwt.t

The current head and all the known (valid) alternate heads.

val mem : Legacy_state.Chain.t -> Tezos_crypto.Block_hash.t -> bool Lwt.t

Test whether a block belongs to the current mainchain.

val set_head : Legacy_state.Chain.t -> Legacy_state.Block.t -> (Legacy_state.Block.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Record a block as the current head of the chain. It returns the previous head.

val test_and_set_head : Legacy_state.Chain.t -> old:Legacy_state.Block.t -> Legacy_state.Block.t -> (bool, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Atomically change the current head of the chain. This returns true whenever the change succeeded, or false when the current head is not equal to the old argument.

val init_head : Legacy_state.Chain.t -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Restores the data about the current head at startup (recomputes the sets of live blocks and operations).