Module Block.Header

Abstract view over block header storage. This module aims to abstract over block header's read, read_opt and known functions by calling the adequate function depending on the block being pruned or not.

val read : (Legacy_store.Block.store * Tezos_crypto.Block_hash.t) -> (Tezos_base.Block_header.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val read_opt : (Legacy_store.Block.store * Tezos_crypto.Block_hash.t) -> Tezos_base.Block_header.t option Lwt.t

val known : (Legacy_store.Block.store * Tezos_crypto.Block_hash.t) -> bool Lwt.t