Module Legacy_store.Block

type store = Chain.store

val get : Chain.store -> store

type contents = {

header : Tezos_base.Block_header.t;

message : string option;

max_operations_ttl : int;

last_allowed_fork_level : Stdlib.Int32.t;

context : Tezos_crypto.Context_hash.t;

metadata : Stdlib.Bytes.t;

}

module Contents : Store_sigs.SINGLE_STORE with type t = store * Tezos_crypto.Block_hash.t and type value := contents

Block header storage used for pruned blocks. Blocks that are not pruned have their header stored in their contents (see Store.Block.Contents). For an abstraction over a block header, please see the State.Block.Header module.

type pruned_contents = {

header : Tezos_base.Block_header.t;

}

module Pruned_contents : Store_sigs.SINGLE_STORE with type t = store * Tezos_crypto.Block_hash.t and type value := pruned_contents

module Block_metadata_hash : Store_sigs.SINGLE_STORE with type t = store * Tezos_crypto.Block_hash.t and type value := Tezos_crypto.Block_metadata_hash.t

module Operation_hashes : Store_sigs.MAP_STORE with type t = store * Tezos_crypto.Block_hash.t and type key = int and type value = Tezos_crypto.Operation_hash.t list

module Operations : Store_sigs.MAP_STORE with type t = store * Tezos_crypto.Block_hash.t and type key = int and type value = Tezos_base.Operation.t list

module Operations_metadata : Store_sigs.MAP_STORE with type t = store * Tezos_crypto.Block_hash.t and type key = int and type value = Stdlib.Bytes.t list

module Operations_metadata_hashes : Store_sigs.MAP_STORE with type t = store * Tezos_crypto.Block_hash.t and type key = int and type value = Tezos_crypto.Operation_metadata_hash.t list

type invalid_block = {

level : int32;

errors : Tezos_error_monad.TzCore.error list;

}

module Invalid_block : Store_sigs.MAP_STORE with type t = store and type key = Tezos_crypto.Block_hash.t and type value = invalid_block

module Predecessors : Store_sigs.MAP_STORE with type t = store * Tezos_crypto.Block_hash.t and type key = int and type value = Tezos_crypto.Block_hash.t

Block predecessors under /blocks/<block_id>/predecessors/<distance>/<block_id>. Used to compute block predecessors in lib_node_shell/state.ml.