Module Block.Header

Abstract view over block header storage. This module aims to abstract over block header's read, read_opt and known functions by calling the adequate function depending on the block being pruned or not.

val read : (Tezos_legacy_store.Legacy_store.Block.store * Tezos_crypto.Block_hash.t) -> Tezos_base__TzPervasives.Block_header.t Tezos_base__TzPervasives.tzresult Lwt.t

val read_opt : (Tezos_legacy_store.Legacy_store.Block.store * Tezos_crypto.Block_hash.t) -> Tezos_base__TzPervasives.Block_header.t option Lwt.t

val known : (Tezos_legacy_store.Legacy_store.Block.store * Tezos_crypto.Block_hash.t) -> bool Lwt.t